Introduce converter function to handle SPDX types.
Some fields in the SPDX model allow for "NOASSERTION" and / or "NONE" values. We use classes for these in the model, but don't want to expose those classes in the API. So, here, we introduce a converter to handle the input values.
Also introduce a new function to format and partially validated package vendors.

Signed-off-by: Anton Bauhofer <anton.bauhofer@tngtech.com>